fix: compatibility with PyGame
Using the library SimpleGUICS2Pygame we fixed the compatibility issues of running these games outside of CodeSkulptor and the SimpleGUITk library. This fix makes sure that the games can be compiled and run outside of that strict environment, on one that emulates it, and therefore makes it possible to run with little or no code change.

Code has been cleaned, commented and verified.